Fixed GC issues.
There were a number of places where unrooted pointers were used, which gave issues with the GC. As part of this change, tracing on the GC has been implemented which can be enabled using a Cargo feature. This generates lots of logging but can be very useful when tracking down GC issues.